requirements.txt
bokeh==0.12.16
gitdb2==2.0.3
GitPython==2.1.10
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.3
packaging==17.1
pandas==0.23.0
Pillow==5.1.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
selenium==3.12.0
six==1.11.0
smmap2==2.0.3
tornado==5.0.2
